requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.9
mysqlclient==2.1.0
paho-mqtt==1.6.1
pycodestyle==2.8.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
